Add hub-specific `Hub.initialize()` and `Hub.finalize()`
`Hub.initialize()` is called right after a hub is connected, similarly,
`Hub.finalize() is called right before a hub is disconnected.

These are to be overriden by user subclasses to provide model-specific
initialization and shutdown logic.

For example, this can be used to calibrate steering and/or to
stop the wehicle.