Fix timing problem when attaching peripherals
Due to a slip in hub connection logic it could happen a peripheral
was deemed as attached even before an 'attach' message has been sent
and received. If in that case hub has been sent a command, it disconnects
(apparrently).

The problem was that peripheral's port not being `None` was used as an
indication whether or not a peripheral is attached (or not). But this is
WRONG, since when peripheral is attached like:

    @attach(CPlusXLMotor, name='front_drive', port=0)
    @attach(CPlusXLMotor, name='rear_drive', port=1)
    class XtremeOffRoader(CPlusHub):

(that is, specifying port explicitly), the port is indeed not `None`!

This commit fixes this by encapsluating the query into new `_is_attached()`
method.

This should fix issue #6.